libs/centrifugo-cpp/services/jwt-generator/requirements.txt
PyJWT==2.8.0
Flask==3.0.0